requirements.txt
streamlit
pandas
plotly
google-generativeai
gspread
google-auth
